NP Migration: Move doc views registry and existing doc views into discover plugin (#53465)
