--- src/legacy/core_plugins/kbn_doc_views/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "kbn_doc_views",
-  "version": "kibana"
-}
