--- src/legacy/core_plugins/kbn_doc_views/public/views/table/table_row_btn_filter_remove.tsx
@@ -1,57 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-import React from 'react';
-import { FormattedMessage } from '@kbn/i18n/react';
-import { EuiToolTip, EuiButtonIcon } from '@elastic/eui';
-import { i18n } from '@kbn/i18n';
-
-export interface Props {
-  onClick: () => void;
-  disabled?: boolean;
-}
-
-export function DocViewTableRowBtnFilterRemove({ onClick, disabled = false }: Props) {
-  const tooltipContent = disabled ? (
-    <FormattedMessage
-      id="kbnDocViews.table.unindexedFieldsCanNotBeSearchedTooltip"
-      defaultMessage="Unindexed fields can not be searched"
-    />
-  ) : (
-    <FormattedMessage
-      id="kbnDocViews.table.filterOutValueButtonTooltip"
-      defaultMessage="Filter out value"
-    />
-  );
-
-  return (
-    <EuiToolTip content={tooltipContent}>
-      <EuiButtonIcon
-        aria-label={i18n.translate('kbnDocViews.table.filterOutValueButtonAriaLabel', {
-          defaultMessage: 'Filter out value',
-        })}
-        className="kbnDocViewer__actionButton"
-        data-test-subj="removeInclusiveFilterButton"
-        disabled={disabled}
-        onClick={onClick}
-        iconType={'magnifyWithMinus'}
-        iconSize={'s'}
-      />
-    </EuiToolTip>
-  );
-}
