+++ src/legacy/core_plugins/kibana/public/discover/build_services.ts
@@ -24,24 +24,27 @@ import {
   ToastsStart,
   IUiSettingsClient,
 } from 'kibana/public';
-import * as docViewsRegistry from 'ui/registry/doc_views';
-import { FilterManager, TimefilterContract, IndexPatternsContract } from 'src/plugins/data/public';
+import {
+  FilterManager,
+  TimefilterContract,
+  IndexPatternsContract,
+  DataPublicPluginStart,
+} from 'src/plugins/data/public';
 import { createSavedSearchesService } from './saved_searches';
-// @ts-ignore
 import { DiscoverStartPlugins } from './plugin';
-import { DataStart } from '../../../data/public';
 import { EuiUtilsStart } from '../../../../../plugins/eui_utils/public';
 import { SharePluginStart } from '../../../../../plugins/share/public';
 import { SavedSearch } from './np_ready/types';
+import { DocViewsRegistry } from './np_ready/doc_views/doc_views_registry';
 
 export interface DiscoverServices {
   addBasePath: (path: string) => string;
   capabilities: Capabilities;
   chrome: ChromeStart;
   core: CoreStart;
-  data: DataStart;
+  data: DataPublicPluginStart;
   docLinks: DocLinksStart;
-  docViewsRegistry: docViewsRegistry.DocViewsRegistry;
+  docViewsRegistry: DocViewsRegistry;
   eui_utils: EuiUtilsStart;
   filterManager: FilterManager;
   indexPatterns: IndexPatternsContract;
@@ -54,7 +57,11 @@ export interface DiscoverServices {
   getSavedSearchUrlById: (id: string) => Promise<string>;
   uiSettings: IUiSettingsClient;
 }
-export async function buildServices(core: CoreStart, plugins: DiscoverStartPlugins) {
+export async function buildServices(
+  core: CoreStart,
+  plugins: DiscoverStartPlugins,
+  docViewsRegistry: DocViewsRegistry
+): Promise<DiscoverServices> {
   const services = {
     savedObjectsClient: core.savedObjects.client,
     indexPatterns: plugins.data.indexPatterns,
