+++ src/legacy/core_plugins/kibana/public/discover/np_ready/components/table/table.tsx
@@ -17,9 +17,9 @@
  * under the License.
  */
 import React, { useState } from 'react';
-import { DocViewRenderProps } from 'ui/registry/doc_views';
 import { DocViewTableRow } from './table_row';
 import { arrayContainsObjects, trimAngularSpan } from './table_helper';
+import { DocViewRenderProps } from '../../doc_views/doc_views_types';
 
 const COLLAPSE_LINE_LENGTH = 350;
