+++ src/legacy/core_plugins/kibana/public/discover/np_ready/components/table/table_helper.test.ts
