+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -1850,6 +1850,25 @@
     "kbn.discover.discoverDescription": "ドキュメントにクエリをかけたりフィルターを適用することで、データをインタラクティブに閲覧できます。",
     "kbn.discover.discoverTitle": "ディスカバー",
     "kbn.discover.documentsAriaLabel": "ドキュメント",
+    "kbn.discover.docViews.table.fieldNamesBeginningWithUnderscoreUnsupportedTooltip": "{underscoreSign} で始まるフィールド名はサポートされていません",
+    "kbn.discover.docViews.table.filterForFieldPresentButtonAriaLabel": "現在のフィールドのフィルター",
+    "kbn.discover.docViews.table.filterForFieldPresentButtonTooltip": "現在のフィールドのフィルター",
+    "kbn.discover.docViews.table.filterForValueButtonAriaLabel": "値でフィルタリング",
+    "kbn.discover.docViews.table.filterForValueButtonTooltip": "値でフィルタリング",
+    "kbn.discover.docViews.table.filterOutValueButtonAriaLabel": "値を除外",
+    "kbn.discover.docViews.table.filterOutValueButtonTooltip": "値を除外します",
+    "kbn.discover.docViews.table.noCachedMappingForThisFieldTooltip": "このフィールドのキャッシュされたマッピングがありません。管理 > インデックスパターンページからフィールドリストを更新してください",
+    "kbn.discover.docViews.table.tableTitle": "表",
+    "kbn.discover.docViews.table.toggleColumnInTableButtonAriaLabel": "表の列を切り替える",
+    "kbn.discover.docViews.table.toggleColumnInTableButtonTooltip": "表の列を切り替える",
+    "kbn.discover.docViews.table.unableToFilterForPresenceOfMetaFieldsTooltip": "メタフィールドの有無でフィルタリングできません",
+    "kbn.discover.docViews.table.unableToFilterForPresenceOfScriptedFieldsTooltip": "スクリプトフィールドの有無でフィルタリングできません",
+    "kbn.discover.docViews.table.unindexedFieldsCanNotBeSearchedTooltip": "インデックスされていないフィールドは検索できません",
+    "kbn.discover.docViews.json.codeEditorAriaLabel": "Elasticsearch ドキュメントの JSON ビューのみを読み込む",
+    "kbn.discover.docViews.json.jsonTitle": "JSON",
+    "kbn.discover.docViews.table.fieldNamesBeginningWithUnderscoreUnsupportedAriaLabel": "警告",
+    "kbn.discover.docViews.table.noCachedMappingForThisFieldAriaLabel": "警告",
+    "kbn.discover.docViews.table.toggleFieldDetails": "フィールド詳細を切り替える",
     "kbn.discover.errorLoadingData": "データの読み込み中にエラーが発生",
     "kbn.discover.fetchError.howToAddressErrorDescription": "このエラーは、{scriptedFields} タブにある {managementLink} の {fetchErrorScript} フィールドを編集することで解決できます。",
     "kbn.discover.fetchError.managmentLinkText": "管理 > インデックスパターン",
@@ -2839,25 +2858,6 @@
     "kbn.advancedSettings.courier.batchSearchesText": "無効の場合、ダッシュボードパネルは個々に読み込まれ、検索リクエストはユーザーが移動するか\n           クエリを更新すると停止します。有効の場合、ダッシュボードパネルはすべてのデータが読み込まれると同時に読み込まれ、\n           検索は停止しません。",
     "kbn.doc.couldNotFindDocumentsDescription": "その ID に一致するドキュメントがありません。",
     "kbn.doc.somethingWentWrongDescription": "{indexName} が欠けています。",
-    "kbnDocViews.table.fieldNamesBeginningWithUnderscoreUnsupportedTooltip": "{underscoreSign} で始まるフィールド名はサポートされていません",
-    "kbnDocViews.table.filterForFieldPresentButtonAriaLabel": "現在のフィールドのフィルター",
-    "kbnDocViews.table.filterForFieldPresentButtonTooltip": "現在のフィールドのフィルター",
-    "kbnDocViews.table.filterForValueButtonAriaLabel": "値でフィルタリング",
-    "kbnDocViews.table.filterForValueButtonTooltip": "値でフィルタリング",
-    "kbnDocViews.table.filterOutValueButtonAriaLabel": "値を除外",
-    "kbnDocViews.table.filterOutValueButtonTooltip": "値を除外します",
-    "kbnDocViews.table.noCachedMappingForThisFieldTooltip": "このフィールドのキャッシュされたマッピングがありません。管理 > インデックスパターンページからフィールドリストを更新してください",
-    "kbnDocViews.table.tableTitle": "表",
-    "kbnDocViews.table.toggleColumnInTableButtonAriaLabel": "表の列を切り替える",
-    "kbnDocViews.table.toggleColumnInTableButtonTooltip": "表の列を切り替える",
-    "kbnDocViews.table.unableToFilterForPresenceOfMetaFieldsTooltip": "メタフィールドの有無でフィルタリングできません",
-    "kbnDocViews.table.unableToFilterForPresenceOfScriptedFieldsTooltip": "スクリプトフィールドの有無でフィルタリングできません",
-    "kbnDocViews.table.unindexedFieldsCanNotBeSearchedTooltip": "インデックスされていないフィールドは検索できません",
-    "kbnDocViews.json.codeEditorAriaLabel": "Elasticsearch ドキュメントの JSON ビューのみを読み込む",
-    "kbnDocViews.json.jsonTitle": "JSON",
-    "kbnDocViews.table.fieldNamesBeginningWithUnderscoreUnsupportedAriaLabel": "警告",
-    "kbnDocViews.table.noCachedMappingForThisFieldAriaLabel": "警告",
-    "kbnDocViews.table.toggleFieldDetails": "フィールド詳細を切り替える",
     "kbnVislibVisTypes.area.areaDescription": "折れ線グラフの下の数量を強調します。",
     "kbnVislibVisTypes.area.areaTitle": "エリア",
     "kbnVislibVisTypes.area.groupTitle": "系列を分割",
