+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -1851,6 +1851,25 @@
     "kbn.discover.discoverDescription": "通过查询和筛选原始文档来以交互方式浏览您的数据。",
     "kbn.discover.discoverTitle": "Discover",
     "kbn.discover.documentsAriaLabel": "文档",
+    "kbn.discover.docViews.table.fieldNamesBeginningWithUnderscoreUnsupportedTooltip": "不支持以 {underscoreSign} 开头的字段名称",
+    "kbn.discover.docViews.table.filterForFieldPresentButtonAriaLabel": "筛留存在的字段",
+    "kbn.discover.docViews.table.filterForFieldPresentButtonTooltip": "筛留存在的字段",
+    "kbn.discover.docViews.table.filterForValueButtonAriaLabel": "筛留值",
+    "kbn.discover.docViews.table.filterForValueButtonTooltip": "筛留值",
+    "kbn.discover.docViews.table.filterOutValueButtonAriaLabel": "筛除值",
+    "kbn.discover.docViews.table.filterOutValueButtonTooltip": "筛除值",
+    "kbn.discover.docViews.table.noCachedMappingForThisFieldTooltip": "此字段没有任何已缓存映射。从“管理”>“索引模式”页面刷新字段列表",
+    "kbn.discover.docViews.table.tableTitle": "表",
+    "kbn.discover.docViews.table.toggleColumnInTableButtonAriaLabel": "在表中切换列",
+    "kbn.discover.docViews.table.toggleColumnInTableButtonTooltip": "在表中切换列",
+    "kbn.discover.docViews.table.unableToFilterForPresenceOfMetaFieldsTooltip": "无法筛留元数据字段",
+    "kbn.discover.docViews.table.unableToFilterForPresenceOfScriptedFieldsTooltip": "无法筛留脚本字段",
+    "kbn.discover.docViews.table.unindexedFieldsCanNotBeSearchedTooltip": "无法搜索未索引字段",
+    "kbn.discover.docViews.json.codeEditorAriaLabel": "Elasticsearch 文档的只读 JSON 视图",
+    "kbn.discover.docViews.json.jsonTitle": "JSON",
+    "kbn.discover.docViews.table.fieldNamesBeginningWithUnderscoreUnsupportedAriaLabel": "警告",
+    "kbn.discover.docViews.table.noCachedMappingForThisFieldAriaLabel": "警告",
+    "kbn.discover.docViews.table.toggleFieldDetails": "切换字段详细信息",
     "kbn.discover.errorLoadingData": "加载数据时出错",
     "kbn.discover.fetchError.howToAddressErrorDescription": "您可以通过编辑 “{scriptedFields}” 选项卡下 “{managementLink}” 中的 “{fetchErrorScript}” 字段来解决此错误。",
     "kbn.discover.fetchError.managmentLinkText": "管理 > 索引模式",
@@ -2840,25 +2859,6 @@
     "kbn.advancedSettings.courier.batchSearchesText": "禁用时，仪表板面板将分别加载，用户离开时或更新查询时，\n           搜索请求将终止。启用时，仪表板面板将一起加载并加载所有数据，\n           搜索将不会终止。",
     "kbn.doc.couldNotFindDocumentsDescription": "无文档匹配该 ID。",
     "kbn.doc.somethingWentWrongDescription": "{indexName} 缺失。",
-    "kbnDocViews.table.fieldNamesBeginningWithUnderscoreUnsupportedTooltip": "不支持以 {underscoreSign} 开头的字段名称",
-    "kbnDocViews.table.filterForFieldPresentButtonAriaLabel": "筛留存在的字段",
-    "kbnDocViews.table.filterForFieldPresentButtonTooltip": "筛留存在的字段",
-    "kbnDocViews.table.filterForValueButtonAriaLabel": "筛留值",
-    "kbnDocViews.table.filterForValueButtonTooltip": "筛留值",
-    "kbnDocViews.table.filterOutValueButtonAriaLabel": "筛除值",
-    "kbnDocViews.table.filterOutValueButtonTooltip": "筛除值",
-    "kbnDocViews.table.noCachedMappingForThisFieldTooltip": "此字段没有任何已缓存映射。从“管理”>“索引模式”页面刷新字段列表",
-    "kbnDocViews.table.tableTitle": "表",
-    "kbnDocViews.table.toggleColumnInTableButtonAriaLabel": "在表中切换列",
-    "kbnDocViews.table.toggleColumnInTableButtonTooltip": "在表中切换列",
-    "kbnDocViews.table.unableToFilterForPresenceOfMetaFieldsTooltip": "无法筛留元数据字段",
-    "kbnDocViews.table.unableToFilterForPresenceOfScriptedFieldsTooltip": "无法筛留脚本字段",
-    "kbnDocViews.table.unindexedFieldsCanNotBeSearchedTooltip": "无法搜索未索引字段",
-    "kbnDocViews.json.codeEditorAriaLabel": "Elasticsearch 文档的只读 JSON 视图",
-    "kbnDocViews.json.jsonTitle": "JSON",
-    "kbnDocViews.table.fieldNamesBeginningWithUnderscoreUnsupportedAriaLabel": "警告",
-    "kbnDocViews.table.noCachedMappingForThisFieldAriaLabel": "警告",
-    "kbnDocViews.table.toggleFieldDetails": "切换字段详细信息",
     "kbnVislibVisTypes.area.areaDescription": "突出折线图下方的数量",
     "kbnVislibVisTypes.area.areaTitle": "面积图",
     "kbnVislibVisTypes.area.groupTitle": "拆分序列",
